// types.ts
export enum PageType {
    COVER = 'COVER',
    INTRO = 'INTRO',
    HISTORY = 'HISTORY',
    ART_STATIC = 'ART_STATIC',
    ART_INTERACTIVE_CALLIGRAPHY = 'ART_INTERACTIVE_CALLIGRAPHY',
    CRAFT_STATIC = 'CRAFT_STATIC',
    CRAFT_INTERACTIVE_REPAIR = 'CRAFT_INTERACTIVE_REPAIR',
    FESTIVAL_STATIC = 'FESTIVAL_STATIC',
    FESTIVAL_INTERACTIVE_LANTERN = 'FESTIVAL_INTERACTIVE_LANTERN',
    CUSTOM_STATIC = 'CUSTOM_STATIC',
    MUSIC_STATIC = 'MUSIC_STATIC',
    MUSIC_INTERACTIVE = 'MUSIC_INTERACTIVE',
    POETRY_INTERACTIVE_AI = 'POETRY_INTERACTIVE_AI',
    MODERN_INHERITANCE = 'MODERN_INHERITANCE',
    CONCLUSION = 'CONCLUSION'
}

export interface PageConfig {
    id: number;
    type: PageType;
    title: string;
    subtitle?: string;
    bgImage?: string;
    content?: string;
}